requirements.txt
asgiref==3.9.1
astroid==3.3.11
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==45.0.7
deepl==1.22.0
dill==0.4.0
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.2
django-mailgun-mime==0.1.7
django-stubs==5.2.5
django-stubs-ext==5.2.5
djangorestframework==3.16.1
djangorestframework-stubs==3.16.2
djangorestframework_simplejwt==5.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-translate==3.21.1
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==6.0.1
mccabe==0.7.0
packaging==25.0
platformdirs==4.4.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pylint==3.3.8
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
python-decouple==3.8
requests==2.32.5
rfc3986==1.5.0
rsa==4.9.1
setuptools==78.1.1
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.3
types-PyYAML==6.0.12.20250822
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
unrpa==2.3.0
urllib3==2.5.0
wheel==0.45.1
